* jest tests * small fixes * pull out the form parsing functions so the pure DOM functions can be   unit tested more easily* jest tests   * small fixes   * pull out the form parsing functions so the pure DOM functions can be     unit tested more easily* jest tests     * small fixes     * pull out the form parsing functions so the pure DOM functions can       be unit tested more easily* jest tests       * small fixes       * pull out the form parsing functions so the pure DOM functions         can be unit tested more easily* jest tests         * small fixes         * pull out the form parsing functions so the pure DOM functions           can be unit tested more easily* jest tests           * small fixes           * pull out the form parsing functions so the pure DOM             functions can be unit tested more easily* jest tests             * small fixes             * pull out the form parsing functions so the pure DOM               functions can be unit tested more easily* jest tests               * small fixes               * pull out the form parsing functions so the pure DOM                 functions can be unit tested more easily
