+++ __tests__/domActions.test.ts
@@ -1,0 +1,7 @@
+import { showExtrinsicStatus } from "../src/domActions";
+import * as assert from "assert";
+import {test} from "@jest/globals";
+
+test("foo", () => {
+    assert(false);
+})
