+++ built/index.js
@@ -6,7 +6,8 @@ import { web3Accounts, web3Enable } from 'https://cdn.jsdelivr.net/npm/@polkadot
 import { Keyring } from 'https://cdn.jsdelivr.net/npm/@polkadot/keyring@12.1.2/+esm';
 // @ts-ignore
 import { options } from "https://cdn.jsdelivr.net/npm/@frequency-chain/api-augment@1.6.1/+esm";
-import { clearFormInvalid, getHTMLInputValue, getSelectedOption, getCreateSponsoredAccountFormData, getGrantDelegationFormData, listenForExtrinsicsChange, setVisibility, validateForm, getApplyItemActionsWithSignatureFormData, getAddPublicKeyFormData, getClaimHandleFormData, getUpsertPageFormData, getDeletePageWithSignatureFormData, setProgress, onProviderEndpointChanged, } from "./domActions.js";
+import { clearFormInvalid, getHTMLInputValue, getSelectedOption, listenForExtrinsicsChange, setVisibility, validateForm, setProgress, onProviderEndpointChanged, } from "./domActions.js";
+import { getCreateSponsoredAccountFormData, getGrantDelegationFormData, getApplyItemActionsWithSignatureFormData, getAddPublicKeyFormData, getClaimHandleFormData, getUpsertPageFormData, getDeletePageWithSignatureFormData, } from "./formApiActions.js";
 import { getBlockNumber, signPayloadWithExtension, signPayloadWithKeyring, submitExtrinsicWithExtension, submitExtrinsicWithKeyring, } from "./chainActions.js";
 // const Hash = interfaces.Hash;
 let PREFIX = 42;
@@ -93,14 +94,14 @@ async function connect(event) {
     setProgress("connect-button", true);
     let selectedProvider = getSelectedOption('provider-list');
     let api;
+    let providerName = "";
     try {
         if (selectedProvider.id === 'other-endpoint-value') {
-            providerName = getHTMLInputValue('other-endpoint-url');
+            providerName = getHTMLInputValue('other-endpoint-url') || "";
             api = await getApi(providerName);
         }
         else {
-            providerName = selectedProvider.getAttribute('name');
-            selectedProvider.getAttribute("name") || "";
+            providerName = selectedProvider.getAttribute("name") || "";
             api = await getApi(selectedProvider.value);
         }
         await updateConnectionStatus(api);
@@ -116,9 +117,8 @@ async function connect(event) {
         alert(`could not connect to ${providerName || "empty value"}. Please enter a valid and reachable Websocket URL.`);
     }
     finally {
-        document.getElementById('connect-button').disabled = false;
+        setProgress("connect-button", false);
     }
-    setProgress("connect-button", false);
     return;
 }
 function populateDropdownWithAccounts(elementId) {
