+++ index.html
@@ -4,7 +4,7 @@
     <title>Frequency Signing dApp</title>
     <link rel="icon" type="image/x-icon" href="./favicon.ico">
     <link rel="stylesheet" href="//unpkg.com/mvp.css@1.12/mvp.css">
-    <link rel="stylesheet" href="/index.css">
+    <link rel="stylesheet" href="./index.css">
 </head>
 <body style="margin: 22px;">
 <div id="connection-info">
@@ -20,7 +20,7 @@
         <option id="other-endpoint-value">Other endpoint</option>
     </select>
 
-    <fieldset id="other-endpoint">
+    <fieldset id="other-endpoint" class="hidden">
         <label for="other-endpoint-url">WSS Endpoint</label>
         <input id="other-endpoint-url" placeholder="wss://some.frequency.node"/>
     </fieldset>
