+++ package.json
@@ -5,8 +5,8 @@
   "main": "index.js",
   "type": "module",
   "scripts": {
-    "build": "tsc ",
-    "test": "nothing to test"
+    "build": "tsc",
+    "test": "jest"
   },
   "repository": {
     "type": "git",
@@ -19,6 +19,16 @@
   },
   "homepage": "https://github.com/LibertyDSNP/signing-ui#readme",
   "devDependencies": {
+    "@babel/preset-env": "^7.22.7",
+    "@babel/preset-typescript": "^7.22.5",
+    "@jest/globals": "^29.6.1",
+    "jest": "^29.6.1",
+    "jest-environment-jsdom": "^29.6.1",
+    "ts-node": "^10.9.1",
     "typescript": "^5.0.4"
+  },
+  "jest": {
+    "bail": false,
+    "clearMocks": true
   }
 }
